requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
numpy==1.22.3
pandas==1.4.2
pip==21.2.4
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
setuptools==61.2.0
six==1.16.0
typing_extensions==4.2.0
wheel==0.37.1
wincertstore==0.2
yarl==1.7.2
youtube-dl==2021.12.17
